fix(d1): bind LIMIT as f64, not i64
D1 errors with D1_TYPE_ERROR: Type 'bigint' not supported for value '<n>'
when an i64 reaches the binding layer. Match the standing wasm/D1 rule:
cast numeric binds to f64.

Verified live: AdminService/ListUsers now returns the demo user
registered through orangevault.

+++ src/services/admin.rs
@@ -42,7 +42,8 @@ impl AdminService for AdminServer {
         _ctx: RequestContext,
         request: OwnedListUsersRequestView,
     ) -> ServiceResult<ListUsersResponse> {
-        let limit = request.limit.unwrap_or(100).clamp(1, 1000) as i64;
+        // D1 rejects i64 — bind LIMIT as f64.
+        let limit = request.limit.unwrap_or(100).clamp(1, 1000) as f64;
 
         let stmt = self
             .db
